+++ CHANGELOG.md
@@ -12,11 +12,18 @@ For a roadmap including expected timeline, please refer to [ROADMAP.md](./ROADMA
 
 ## [0.4.0]
 
+### Added
+
 - ADDED: Clear statement that `ElementRef` referenced element MUST exist locally in the same entity.
 - ADDED: `meta.document.title` to give the overall document a human-readable title
+-
+
+### Changed
+
 - BREAKING: wrong @Semantics.amountCurrencyCode annotation, should be @Semantics.amount.currencyCode
 - BREAKING: wrong @Semantics.quantityUnitOfMeasure annotation, should be @Semantics.quantity.unitOfMeasure
 - BREAKING: For custom types, we should not set `key` property - as this is decided on entity element level.
+- FIXED: JSON Schema export $ref from annotation extensions back to core spec (`ElementReference`) was missing.
 
 ## [0.3.0]